Add migration Orders

+++ database/migrations/2018_03_07_144851_create_products_table.php
@@ -18,7 +18,7 @@ class CreateProductsTable extends Migration
             $table->tinyInteger('seller');
             $table->string('title');
             $table->tinyInteger('count');
-            $table->tinyInteger('price');
+            $table->double('price', 8, 2);
             $table->timestamps();
         });
     }
